data/docs/module3/api-vetements-ia/requirements.txt
# Dépendances principales
Flask==2.2.3
Pillow==9.4.0
numpy==1.24.2
tensorflow==2.12.0
gunicorn==20.1.0

# Utilitaires
python-dotenv==1.0.0
requests==2.28.2
flask-cors==3.0.10

# Optimisation
tensorflow-model-optimization==0.7.3

# Développement
pytest==7.3.1
black==23.3.0
flake8==6.0.0